cfx_statedb

Struct StateDbGeneric

pub struct StateDbGeneric { /* private fields */ }

Implementations

impl StateDb

pub fn new(storage: Box<dyn StorageStateTrait>) -> Self

pub fn delete( &mut self, key: StorageKeyWithSpace<'_>, debug_record: Option<&mut ComputeEpochDebugRecord>, ) -> Result<()>

pub fn delete_all<AM: AccessMode>( &mut self, key_prefix: StorageKeyWithSpace<'_>, debug_record: Option<&mut ComputeEpochDebugRecord>, ) -> Result<Vec<MptKeyValue>>

pub fn set_storage_layout( &mut self, address: &AddressWithSpace, storage_layout: StorageLayout, debug_record: Option<&mut ComputeEpochDebugRecord>, ) -> Result<()>

pub fn compute_state_root( &mut self, debug_record: Option<&mut ComputeEpochDebugRecord>, ) -> Result<StateRootWithAuxInfo>

This method is only used for genesis block because state root is required to compute genesis epoch_id. For other blocks there are deferred execution so the state root computation is merged inside commit method.

pub fn commit( &mut self, epoch_id: EpochId, debug_record: Option<&mut ComputeEpochDebugRecord>, ) -> Result<StateRootWithAuxInfo>