cfx_statedb::global_params

Trait GlobalParamKey

pub trait GlobalParamKey {
    const ID: usize;
    const KEY: &'static [u8];
    const SPACE: Space = Space::Native;
    const ADDRESS: Address = STORAGE_INTEREST_STAKING_CONTRACT_ADDRESS;
    const STORAGE_KEY: StorageKeyWithSpace<'static> = _;

    // Provided methods
    fn init_vm_value() -> U256 { ... }
    fn from_vm_value(val: U256) -> U256 { ... }
    fn into_vm_value(val: U256) -> U256 { ... }
}

Required Associated Constants

const ID: usize

const KEY: &'static [u8]

Provided Associated Constants

const SPACE: Space = Space::Native

const ADDRESS: Address = STORAGE_INTEREST_STAKING_CONTRACT_ADDRESS

const STORAGE_KEY: StorageKeyWithSpace<'static> = _

Provided Methods

fn init_vm_value() -> U256

How to initialize such a variable in the executor

fn from_vm_value(val: U256) -> U256

How to convert such a variable from the executor representing to the db representing

fn into_vm_value(val: U256) -> U256

How to convert such a variable from the db representing to the executor representing

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.